feat: added validation for current functions.

+++ src/main/java/com/example/demo/service/TraineeService.java
@@ -2,7 +2,9 @@ package com.example.demo.service;
 
 import com.example.demo.repository.TraineeRepository;
 import com.example.demo.domain.Trainee;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 public class TraineeService {
@@ -17,6 +19,9 @@ public class TraineeService {
     }
 
     public void deleteTrainee(long id) {
+        if (!traineeRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, " no such trainee");
+        }
         traineeRepository.deleteById(id);
     }
