+++ src/main/java/com/example/demo/service/TrainerService.java
@@ -4,7 +4,9 @@ import com.example.demo.domain.Trainee;
 import com.example.demo.domain.Trainer;
 import com.example.demo.repository.TraineeRepository;
 import com.example.demo.repository.TrainerRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 public class TrainerService {
@@ -19,6 +21,9 @@ public class TrainerService {
     }
 
     public void deleteTrainer(long id) {
+        if (!trainerRepository.existsById(id)) {
+            throw new  ResponseStatusException(HttpStatus.NOT_FOUND, " no such trainer");
+        }
         trainerRepository.deleteById(id);
     }
 }
